assure that invoice can only be generated once
